materialize-snowflake: move authentication into `go/auth/snowflake` package
Refactoring to move Snowflake authentication logic into a dedicated
package under `go/auth/snowflake`. No functional changes, just
re-organization to avoid duplicate code since `source-snowflake` will
be updated to use the same authentication logic in the next commit.
