+++ go/auth/snowflake/.snapshots/TestConfigSchema
@@ -1,0 +1,71 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "https://github.com/estuary/connectors/go/auth/snowflake/credential-config",
+  "oneOf": [
+    {
+      "properties": {
+        "auth_type": {
+          "type": "string",
+          "const": "jwt",
+          "default": "jwt"
+        },
+        "user": {
+          "type": "string",
+          "title": "User",
+          "description": "The Snowflake user login name",
+          "order": 1
+        },
+        "private_key": {
+          "type": "string",
+          "title": "Private Key",
+          "description": "Private Key to be used to sign the JWT token",
+          "multiline": true,
+          "order": 2,
+          "secret": true
+        }
+      },
+      "required": [
+        "auth_type",
+        "private_key"
+      ],
+      "title": "Private Key (JWT)"
+    },
+    {
+      "properties": {
+        "auth_type": {
+          "type": "string",
+          "const": "user_password",
+          "default": "user_password"
+        },
+        "user": {
+          "type": "string",
+          "title": "User",
+          "description": "The Snowflake user login name",
+          "order": 1
+        },
+        "password": {
+          "type": "string",
+          "title": "Password",
+          "description": "The password for the provided user",
+          "order": 2,
+          "secret": true
+        }
+      },
+      "required": [
+        "auth_type",
+        "user",
+        "password"
+      ],
+      "title": "User Password"
+    }
+  ],
+  "type": "object",
+  "title": "Test Config Schema",
+  "description": "Snowflake Credentials",
+  "default": {
+    "auth_type": "jwt"
+  },
+  "discriminator": {
+    "propertyName": "auth_type"
+  }
+}
