+++ materialize-snowflake/config.go
@@ -1,20 +1,17 @@
 package main
 
 import (
-	"crypto/rsa"
 	"crypto/x509"
 	"encoding/base64"
-	"encoding/pem"
 	"fmt"
 	"net/url"
 	"regexp"
 	"strings"
 
+	snowflake_auth "github.com/estuary/connectors/go/auth/snowflake"
 	"github.com/estuary/connectors/go/dbt"
 	boilerplate "github.com/estuary/connectors/materialize-boilerplate"
-	"github.com/invopop/jsonschema"
 	sf "github.com/snowflakedb/gosnowflake"
-	orderedmap "github.com/wk8/go-ordered-map/v2"
 )
 
 var featureFlagDefaults = map[string]bool{
@@ -24,18 +21,17 @@ var featureFlagDefaults = map[string]bool{
 }
 
 type config struct {
-	Host          string                     `json:"host" jsonschema:"title=Host (Account URL),description=The Snowflake Host used for the connection. Must include the account identifier and end in .snowflakecomputing.com. Example: orgname-accountname.snowflakecomputing.com (do not include the protocol)." jsonschema_extras:"order=0,pattern=^[^/:]+.snowflakecomputing.com$"`
-	Database      string                     `json:"database" jsonschema:"title=Database,description=The SQL database to connect to." jsonschema_extras:"order=3"`
-	Schema        string                     `json:"schema" jsonschema:"title=Schema,description=Database schema for bound collection tables (unless overridden within the binding resource configuration)." jsonschema_extras:"order=4"`
-	Warehouse     string                     `json:"warehouse,omitempty" jsonschema:"title=Warehouse,description=The Snowflake virtual warehouse used to execute queries. Uses the default warehouse for the Snowflake user if left blank." jsonschema_extras:"order=5"`
-	Role          string                     `json:"role,omitempty" jsonschema:"title=Role,description=The user role used to perform actions." jsonschema_extras:"order=6"`
-	Account       string                     `json:"account,omitempty" jsonschema:"title=Account,description=Optional Snowflake account identifier." jsonschema_extras:"order=7"`
-	HardDelete    bool                       `json:"hardDelete,omitempty" jsonschema:"title=Hard Delete,description=If this option is enabled items deleted in the source will also be deleted from the destination. By default is disabled and _meta/op in the destination will signify whether rows have been deleted (soft-delete).,default=false" jsonschema_extras:"order=8"`
-	Credentials   credentialConfig           `json:"credentials" jsonschema:"title=Authentication"`
-	Schedule      boilerplate.ScheduleConfig `json:"syncSchedule,omitempty" jsonschema:"title=Sync Schedule,description=Configure schedule of transactions for the materialization."`
-	DBTJobTrigger dbt.JobConfig              `json:"dbt_job_trigger,omitempty" jsonschema:"title=dbt Cloud Job Trigger,description=Trigger a dbt Job when new data is available"`
-
-	Advanced advancedConfig `json:"advanced,omitempty" jsonschema:"title=Advanced Options,description=Options for advanced users. You should not typically need to modify these." jsonschema_extra:"advanced=true"`
+	Host          string                           `json:"host" jsonschema:"title=Host (Account URL),description=The Snowflake Host used for the connection. Must include the account identifier and end in .snowflakecomputing.com. Example: orgname-accountname.snowflakecomputing.com (do not include the protocol)." jsonschema_extras:"order=0,pattern=^[^/:]+.snowflakecomputing.com$"`
+	Database      string                           `json:"database" jsonschema:"title=Database,description=The SQL database to connect to." jsonschema_extras:"order=3"`
+	Schema        string                           `json:"schema" jsonschema:"title=Schema,description=Database schema for bound collection tables (unless overridden within the binding resource configuration)." jsonschema_extras:"order=4"`
+	Warehouse     string                           `json:"warehouse,omitempty" jsonschema:"title=Warehouse,description=The Snowflake virtual warehouse used to execute queries. Uses the default warehouse for the Snowflake user if left blank." jsonschema_extras:"order=5"`
+	Role          string                           `json:"role,omitempty" jsonschema:"title=Role,description=The user role used to perform actions." jsonschema_extras:"order=6"`
+	Account       string                           `json:"account,omitempty" jsonschema:"title=Account,description=Optional Snowflake account identifier." jsonschema_extras:"order=7"`
+	HardDelete    bool                             `json:"hardDelete,omitempty" jsonschema:"title=Hard Delete,description=If this option is enabled items deleted in the source will also be deleted from the destination. By default is disabled and _meta/op in the destination will signify whether rows have been deleted (soft-delete).,default=false" jsonschema_extras:"order=8"`
+	Credentials   *snowflake_auth.CredentialConfig `json:"credentials" jsonschema:"title=Authentication"`
+	Schedule      boilerplate.ScheduleConfig       `json:"syncSchedule,omitempty" jsonschema:"title=Sync Schedule,description=Configure schedule of transactions for the materialization."`
+	DBTJobTrigger dbt.JobConfig                    `json:"dbt_job_trigger,omitempty" jsonschema:"title=dbt Cloud Job Trigger,description=Trigger a dbt Job when new data is available"`
+	Advanced      advancedConfig                   `json:"advanced,omitempty" jsonschema:"title=Advanced Options,description=Options for advanced users. You should not typically need to modify these." jsonschema_extra:"advanced=true"`
 }
 
 type advancedConfig struct {
@@ -80,12 +76,12 @@ func (c *config) toURI(tenant string) (string, error) {
 
 	// Authentication
 	var user string
-	if c.Credentials.AuthType == UserPass {
+	if c.Credentials.AuthType == snowflake_auth.UserPass {
 		user = url.QueryEscape(c.Credentials.User) + ":" + url.QueryEscape(c.Credentials.Password)
-	} else if c.Credentials.AuthType == JWT {
+	} else if c.Credentials.AuthType == snowflake_auth.JWT {
 		// We run this as part of validate to ensure that there is no error, so
 		// this is not expected to error here.
-		if key, err := c.Credentials.privateKey(); err != nil {
+		if key, err := c.Credentials.ParsePrivateKey(); err != nil {
 			return "", err
 		} else if privateKey, err := x509.MarshalPKCS8PrivateKey(key); err != nil {
 			return "", fmt.Errorf("parsing private key: %w", err)
@@ -101,24 +97,6 @@ func (c *config) toURI(tenant string) (string, error) {
 
 	dsn := user + "@" + uri.Hostname() + ":" + uri.Port() + "?" + queryParams.Encode()
 	return dsn, nil
-}
-
-func (c *credentialConfig) privateKey() (*rsa.PrivateKey, error) {
-	if c.AuthType == JWT {
-		// When providing the PEM file in a JSON file, newlines can't be specified unless
-		// escaped, so here we allow an escape hatch to parse these PEM files
-		var pkString = strings.ReplaceAll(c.PrivateKey, "\\n", "\n")
-		var block, _ = pem.Decode([]byte(pkString))
-		if block == nil {
-			return nil, fmt.Errorf("invalid private key: must be PEM format")
-		} else if key, err := x509.ParsePKCS8PrivateKey(block.Bytes); err != nil {
-			return nil, fmt.Errorf("parsing private key: %w", err)
-		} else {
-			return key.(*rsa.PrivateKey), nil
-		}
-	}
-
-	return nil, fmt.Errorf("only supported with JWT authentication")
 }
 
 var hostRe = regexp.MustCompile(`(?i)^.+.snowflakecomputing\.com$`)
@@ -163,129 +141,4 @@ func (c *config) Validate() error {
 	}
 
 	return validHost(c.Host)
-}
-
-const (
-	UserPass = "user_password" // username and password
-	JWT      = "jwt"           // JWT, needs a private key
-)
-
-type credentialConfig struct {
-	AuthType   string `json:"auth_type"`
-	User       string `json:"user"`
-	Password   string `json:"password"`
-	PrivateKey string `json:"private_key"`
-}
-
-func (c *credentialConfig) Validate() error {
-	switch c.AuthType {
-	case UserPass:
-		return c.validateUserPassCreds()
-	case JWT:
-		return c.validateJWTCreds()
-	default:
-		return fmt.Errorf("invalid credentials auth type %q", c.AuthType)
-	}
-}
-
-func (c *credentialConfig) validateUserPassCreds() error {
-	if c.User == "" {
-		return fmt.Errorf("missing user")
-	}
-	if c.Password == "" {
-		return fmt.Errorf("missing password")
-	}
-
-	return nil
-}
-
-func (c *credentialConfig) validateJWTCreds() error {
-	if c.User == "" {
-		return fmt.Errorf("missing user")
-	}
-	if c.PrivateKey == "" {
-		return fmt.Errorf("missing private_key")
-	}
-
-	if _, err := c.privateKey(); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// JSONSchema allows for the schema to be (semi-)manually specified when used with the
-// github.com/invopop/jsonschema package in go-schema-gen, to fullfill the required schema shape for
-// our oauth
-func (credentialConfig) JSONSchema() *jsonschema.Schema {
-	uProps := orderedmap.New[string, *jsonschema.Schema]()
-	uProps.Set("auth_type", &jsonschema.Schema{
-		Type:    "string",
-		Default: UserPass,
-		Const:   UserPass,
-	})
-	uProps.Set("user", &jsonschema.Schema{
-		Title:       "User",
-		Description: "The Snowflake user login name",
-		Type:        "string",
-		Extras: map[string]interface{}{
-			"order": 1,
-		},
-	})
-	uProps.Set("password", &jsonschema.Schema{
-		Title:       "Password",
-		Description: "The password for the provided user",
-		Type:        "string",
-		Extras: map[string]interface{}{
-			"secret": true,
-			"order":  2,
-		},
-	})
-
-	jwtProps := orderedmap.New[string, *jsonschema.Schema]()
-	jwtProps.Set("auth_type", &jsonschema.Schema{
-		Type:    "string",
-		Default: JWT,
-		Const:   JWT,
-	})
-	jwtProps.Set("user", &jsonschema.Schema{
-		Title:       "User",
-		Description: "The Snowflake user login name",
-		Type:        "string",
-		Extras: map[string]interface{}{
-			"order": 1,
-		},
-	})
-	jwtProps.Set("private_key", &jsonschema.Schema{
-		Title:       "Private Key",
-		Description: "Private Key to be used to sign the JWT token",
-		Type:        "string",
-		Extras: map[string]interface{}{
-			"secret":    true,
-			"multiline": true,
-			"order":     2,
-		},
-	})
-
-	return &jsonschema.Schema{
-		Title:       "Authentication",
-		Description: "Snowflake Credentials",
-		Default:     map[string]string{"auth_type": JWT},
-		OneOf: []*jsonschema.Schema{
-			{
-				Title:      "Private Key (JWT)",
-				Required:   []string{"auth_type", "private_key"},
-				Properties: jwtProps,
-			},
-			{
-				Title:      "User Password",
-				Required:   []string{"auth_type", "user", "password"},
-				Properties: uProps,
-			},
-		},
-		Extras: map[string]interface{}{
-			"discriminator": map[string]string{"propertyName": "auth_type"},
-		},
-		Type: "object",
-	}
 }
